Add .haml to suffixesadd
It's handy to be able to quickly
open haml views.

Adding this in vimrc wasn't working, though.
Spelunking revealed that [vim-ruby][]
will use setlocal to change the value of
suffixesadd when editing a ruby file.

So I've added a ruby `after/ftplugin`
to restore the global value set by vimrc.

Thanks to @thorncp for digging up the
right setlocal usage.

  [vim-ruby]: https://github.com/vim-ruby/vim-ruby/blob/ef2deeae81f540a17a3da5907146b3579c8ee0d6/ftplugin/ruby.vim#L48